refactoring over algorithms interface
