+++ src/esbmc/CMakeLists.txt
@@ -25,8 +25,9 @@ target_include_directories(esbmc
     PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}
     PRIVATE ${Boost_INCLUDE_DIRS}
 )
+
 target_link_libraries(esbmc ${OLD_FRONTEND_TARGETS} clangcfrontend 
   clangcppfrontend symex pointeranalysis langapi util_esbmc bigint 
-  solvers clibs default_message goto_algorithms ${Boost_LIBRARIES})
+  solvers clibs default_message gotoalgorithms ${Boost_LIBRARIES})
 
 install(TARGETS esbmc DESTINATION bin)
