+++ src/util/CMakeLists.txt
@@ -1,3 +1,4 @@
+add_library(algorithms algorithms.cpp)
 add_library(util_esbmc irep2.cpp xml_irep.cpp xml.cpp
         arith_tools.cpp base_type.cpp cmdline.cpp config.cpp context.cpp
         expr_util.cpp i2string.cpp location.cpp
@@ -24,3 +25,4 @@ target_include_directories(util_esbmc
 target_link_libraries(util_esbmc default_message fmt::fmt ${Boost_LIBRARIES})
 
 add_subdirectory(message)
+target_link_libraries(algorithms gotoprograms)
